deleted duplicate code and added drawutils
