+++ CMakeLists.txt
@@ -70,6 +70,7 @@ set(SOURCES
 	nvg.cpp
 	glfw.cpp
 	boardview.cpp
+	drawutils.cpp
 	menu.cpp
 	achievement.cpp
 	scoremanager.cpp
