+++ drawutils.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "nvg.h"
+
+void drawShadow(NVGcontext* context, Rect rect);
